requirements.txt
Flask
pytesseract
Pillow
fpdf
reportlab

